+++ lib/src/components/processors/raw_data_processor.dart
@@ -1,3 +1,5 @@
+import 'dart:typed_data';
+
 import '../../models/session.dart';
 import '../../utils/utils.dart';
 
@@ -6,5 +8,6 @@ import 'session_processor.dart';
 /// Processes sessions into raw json format
 class RawDataProcessor extends SessionProcessor {
   @override
-  Future process(Session session) => exportNumericData(session.toJson());
+  Future<Uint8List?> process(Session session) =>
+      exportNumericData(session.toJson());
 }