+++ lib/src/utils/utils.dart
@@ -9,8 +9,8 @@ Future<Uint8List> exportHeatMap(Image image) async {
 }
 
 /// Serializes a json map into a formatted string.
-Future exportNumericData(Map<String, dynamic> data) async =>
-    JsonEncoder.withIndent('\t').convert(data);
+Future<Uint8List> exportNumericData(Map<String, dynamic> data) async =>
+    utf8.encoder.convert(JsonEncoder.withIndent('\t').convert(data));
 
 /// Provides a current timestamp in milliseconds.
 int getTimestamp() => DateTime.now().millisecondsSinceEpoch;